[SofaBaseCollision] Extract the narrow phase code from BruteForceDetection into a new component
