+++ SofaKernel/modules/SofaBaseCollision/CMakeLists.txt
@@ -12,6 +12,7 @@ set(HEADER_FILES
     ${SOFABASECOLLISION_SRC}/BaseProximityIntersection.h
     ${SOFABASECOLLISION_SRC}/BruteForceBroadPhase.h
     ${SOFABASECOLLISION_SRC}/BruteForceDetection.h
+    ${SOFABASECOLLISION_SRC}/BVHNarrowPhase.h
     ${SOFABASECOLLISION_SRC}/CapsuleIntTool.h
     ${SOFABASECOLLISION_SRC}/CapsuleIntTool.inl
     ${SOFABASECOLLISION_SRC}/CapsuleModel.h
@@ -54,6 +55,7 @@ set(SOURCE_FILES
     ${SOFABASECOLLISION_SRC}/BaseProximityIntersection.cpp
     ${SOFABASECOLLISION_SRC}/BruteForceBroadPhase.cpp
     ${SOFABASECOLLISION_SRC}/BruteForceDetection.cpp
+    ${SOFABASECOLLISION_SRC}/BVHNarrowPhase.cpp
     ${SOFABASECOLLISION_SRC}/CapsuleIntTool.cpp
     ${SOFABASECOLLISION_SRC}/CapsuleModel.cpp
     ${SOFABASECOLLISION_SRC}/ContactListener.cpp
