+++ SofaKernel/modules/SofaBaseCollision/src/SofaBaseCollision/BruteForceDetection.h
@@ -23,133 +23,26 @@
 #include <SofaBaseCollision/config.h>
 
 #include <SofaBaseCollision/BruteForceBroadPhase.h>
-#include <sofa/core/collision/NarrowPhaseDetection.h>
+#include <SofaBaseCollision/BVHNarrowPhase.h>
 #include <SofaBaseCollision/CubeModel.h>
 #include <sofa/helper/vector.h>
-
-#include <queue>
-#include <stack>
-
-namespace sofa::core::collision
-{
-    class ElementIntersector;
-}
 
 namespace sofa::component::collision
 {
 
-class MirrorIntersector;
-
 class SOFA_SOFABASECOLLISION_API BruteForceDetection :
     public BruteForceBroadPhase,
-    public core::collision::NarrowPhaseDetection
+    public BVHNarrowPhase
 {
-    /// Range defined by two iterators in a container of CollisionElement
-    using CollisionIteratorRange = std::pair<core::CollisionElementIterator, core::CollisionElementIterator>;
-
-    /// A pair of CollisionIteratorRange where the first range contains elements from a first collision model
-    /// and the second range contains collision elements from a second collision model
-    /// This type is used when testing collision between two collision models
-    /// Note that the second collision model can be the same than the first in case of self collision
-    using TestPair = std::pair< CollisionIteratorRange, CollisionIteratorRange >;
 
 public:
-    SOFA_CLASS2(BruteForceDetection, BruteForceBroadPhase, core::collision::NarrowPhaseDetection);
+    SOFA_CLASS2(BruteForceDetection, BruteForceBroadPhase, BVHNarrowPhase);
 
 protected:
     BruteForceDetection();
 
     ~BruteForceDetection() override = default;
 
-public:
-
-    //////////////////////////////
-    /// NARROW PHASE interface ///
-    //////////////////////////////
-
-    /** \brief In the narrow phase, examine a potential collision between a pair of collision models, which has
-     * been detected in the broad phase.
-     *
-     * The function traverses the hierarchy of CollisionElement's contained in each CollisionModel to avoid
-     * unnecessary pair intersections.
-     * An iterative form of the hierarchy traversal is adopted instead of a recursive form.
-     */
-    void addCollisionPair(const std::pair<core::CollisionModel*, core::CollisionModel*>& cmPair) override;
-
-
-    void draw(const core::visual::VisualParams* /* vparams */) override { }
-
-    inline bool needsDeepBoundingTree() const override { return true; }
-
-protected:
-
-    /// Return true if both collision models belong to the same object, false otherwise
-    static bool isSelfCollision(core::CollisionModel* cm1, core::CollisionModel* cm2);
-
-    /// Build a list of TestPair's from internal and external children of two CollisionModel's
-    static void initializeExternalCells(
-            core::CollisionModel *cm1,
-            core::CollisionModel *cm2,
-            std::queue<TestPair>& externalCells);
-
-    /// Store data related to two finest CollisionModel's
-    struct FinestCollision
-    {
-        core::CollisionModel* cm1 { nullptr };
-        core::CollisionModel* cm2 { nullptr };
-
-        /// ElementIntersector corresponding to cm1 and cm2
-        core::collision::ElementIntersector* intersector { nullptr };
-
-        // True in case cm1 and cm2 belong to the same object, false otherwise
-        bool selfCollision { false };
-    };
-
-    void processExternalCell(const TestPair &externalCell,
-                             core::CollisionModel *&cm1,
-                             core::CollisionModel *&cm2,
-                             core::collision::ElementIntersector *coarseIntersector,
-                             const FinestCollision &finest,
-                             MirrorIntersector *mirror,
-                             std::queue<TestPair> &externalCells,
-                             sofa::core::collision::DetectionOutputVector *&outputs) const;
-
-    static void
-    processInternalCell(const TestPair &internalCell,
-                        core::collision::ElementIntersector *coarseIntersector,
-                        const FinestCollision &finest,
-                        std::queue<TestPair> &externalCells,
-                        std::stack<TestPair> &internalCells,
-                        sofa::core::collision::DetectionOutputVector *&outputs);
-
-    static void visitCollisionElements(const TestPair &root,
-                                       core::collision::ElementIntersector *coarseIntersector,
-                                       const FinestCollision &finest,
-                                       std::queue<TestPair> &externalCells,
-                                       std::stack<TestPair> &internalCells,
-                                       sofa::core::collision::DetectionOutputVector *&outputs);
-
-    static void
-    visitExternalChildren(const core::CollisionElementIterator &it1, const core::CollisionElementIterator &it2,
-                          core::collision::ElementIntersector *coarseIntersector,
-                          const FinestCollision &finest,
-                          std::queue<TestPair> &externalCells,
-                          sofa::core::collision::DetectionOutputVector *&outputs);
-
-    /// Test intersection between two ranges of CollisionElement's
-    /// The provided TestPair contains ranges of external CollisionElement's, which means that
-    /// they can be tested against each other for intersection
-    static void finalCollisionPairs(const TestPair& pair,
-                             bool selfCollision,
-                             core::collision::ElementIntersector* intersector,
-                             sofa::core::collision::DetectionOutputVector*& outputs);
-
-private:
-
-    /// Get both collision models corresponding to the provided TestPair
-    static std::pair<core::CollisionModel*, core::CollisionModel*> getCollisionModelsFromTestPair(const TestPair& pair);
-
-    static bool isRangeEmpty(const CollisionIteratorRange& range);
 };
 
 } // namespace sofa::component::collision
